refactor: rename `ensureSocket` to `send`, moving into `Request` module
